+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 # Release Notes
 
+## 1.20
+- Experimental version of the download-agent. Trying to reduce download failures at the beginning of a job. For internal use only.
+
 ## 1.19
 - Bug fix for DxNI error
 - Limit the size of the name of jobs in a scatter